+++ plugins/org.ietr.preesm.pimm.algorithm/src/org/ietr/preesm/pimm/algorithm/spider/codegen/visitor/SpiderCodegenVisitor.java
@@ -69,6 +69,7 @@ import org.ietr.preesm.experiment.model.pimm.DataOutputInterface;
 import org.ietr.preesm.experiment.model.pimm.DataOutputPort;
 import org.ietr.preesm.experiment.model.pimm.DataPort;
 import org.ietr.preesm.experiment.model.pimm.Delay;
+import org.ietr.preesm.experiment.model.pimm.DelayActor;
 import org.ietr.preesm.experiment.model.pimm.Dependency;
 import org.ietr.preesm.experiment.model.pimm.ExecutableActor;
 import org.ietr.preesm.experiment.model.pimm.Expression;
@@ -786,10 +787,10 @@ public class SpiderCodegenVisitor extends PiMMSwitch<Boolean> {
     throw new UnsupportedOperationException();
   }
 
-  // @Override
-  // public Boolean caseConfigInputInterface(final ConfigInputInterface cii) {
-  // throw new UnsupportedOperationException();
-  // }
+  @Override
+  public Boolean caseDelayActor(final DelayActor da) {
+    return true;
+  }
 
   @Override
   public Boolean caseConfigInputPort(final ConfigInputPort cip) {
@@ -883,5 +884,5 @@ public class SpiderCodegenVisitor extends PiMMSwitch<Boolean> {
   // this.method = method;
   // }
   //
-  // }
+  // }w
 }